pyproject.toml
[project]
name = "million-songs-pipeline"
version = "0.1.0"
requires-python = ">=3.11,<3.14"
dependencies = []

[dependency-groups]
jobs = [
    "google-cloud-storage>=2.4,<3.2",
    "pyarrow>=23.0.1",
    "h5py>=3.16.0",
    "requests>=2",
]
local = [
    "prefect==3.6.0",
    "python-dotenv>=1.2.2",
    "google-cloud-run>=0.10.0",
    "google-cloud-bigquery>=3.41.0",
    "db-dtypes>=1.4.2",
    "dbt-bigquery>=1.11.1",
    "plotly>=6.7.0",
    "tidepool-theme>=0.2.0",
    "streamlit>=1.56.0",
    "reverse-geocoder>=1.5.1",
    "griffe<1.15",
]
dev = [
    "jupytext>=1.19.1",
    "notebook>=7.5.5",
]
